+++ src/window.rs
@@ -764,6 +764,21 @@ impl Window {
     pub fn set_cursor_visible(&self, visible: bool) {
         self.window.set_cursor_visible(visible)
     }
+
+    /// Moves the window with the left mouse button until the button is released.
+    ///
+    /// There's no guarantee that this will work unless the left mouse button was pressed
+    /// immediately before this function is called.
+    ///
+    /// ## Platform-specific
+    ///
+    /// - **X11:** Un-grabs the cursor.
+    /// - **macOS:** May prevent the button release event to be triggered.
+    /// - **iOS / Android / Web:** Always returns an [`ExternalError::NotSupported`].
+    #[inline]
+    pub fn set_drag_window(&self) -> Result<(), ExternalError> {
+        self.window.set_drag_window()
+    }
 }
 
 /// Monitor info functions.